Add multiple, optionlist, serializtion for selects
Add the ability to specify multiple options in a `select` column with
`multiple:true`. When the user can specify multiple values we serialize
the values into a single string as CSV by default, though
`serialization:json` causes it to be serialized as a JSON array.

Add the ability to specify a simple comma-separated list of options
without using an external tab in a select column with
`optionlist:a,b,c`. When specifying options this way the field is
displayed as a group of checkboxes/radio buttons depending on the value
of `multiple`.