Add search view to frontend
Add a button to get to a search window if at least one column has search
enabled. The search window will give you a global input for each search
field; when you enter values and hit "Search" you will get back a set of
results that match the entered values _exactly_. Note that this includes
free text inputs; the string values must match exactly, case sensitive.
The results will be rendered as read-only instances of the form with all
fields filled in as they were when the form was submitted.