Enforce a boolean value for boolean inputs
There is no sensical role for an empty value in a boolean field, so we
should ensure that a boolean input is always submitted as either `true`
or `false`. If a form requires an additional "empty" value (e.g., "this
field should be reset to neither `true` nor `false`"), then the field
should be of a type that actually supports those additional values (for
example, a select field).