Rolling refresh of authentication session
Expect authentication plugins to set an `expiry_date` in the
authentication session token. When we reach that time we will treat the
authentication as "stale", meaning we won't use it for authorization,
but we will keep it around until we have the opportunity to refresh it
(which we can only do during a non-API call). When we refresh the
authentication, we pass off to the proper plugin (we expect the plugin
to set an `issuer` field in the authentication session token to allow
this) and have it either refresh the authentication lease or allow the
lease to expire.

The Google OAuth authentication plugin uses the user's Google access
token to retrieve information about the user's profile to establish that
the authentication is valid. In order to do this we have to store the
access token itself in the authentication session token for use during
future transactions.