Add `layout` array to schema for non-input content
Support a new type of configuration row called `text` that takes whatever text is in the second column, parses it as markdown, sanitizes the resulting HTML, and then renders it on the page interleaved with form inputs. Note that any columns that are part of the `index` still get hoisted to the top of the form.

Parse chatter as markdown and render it as HTML. Chatter is still rendered before the index entry, so it can be used as free-form text content before the index entry fields.