Avoid breaking email equality check when undefined
This implicitly had been assuming that if an `anyoneWithLink` permission
were present, it would be the first entry in the
`filePermissions.data.permissions` array. (That check function would
return `true` as a result, terminating the `.some` call early.)

It turns out that isn't necessarily true, which was causing this
function to throw a `TypeError`. Because `emailCanRead` in turn was
throwing an exception rather than returning a boolean (let alone the
necessary `true`), this meant the resulting behavior was that the form
in question _did_ require a login in order to work, despite the desired
behavior being the opposite.

(If the user _was_ logged in, then `email` never was undefined, and that
user would be granted access.)